+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/profiler/GcEventMonitor.java
@@ -51,7 +51,7 @@ public class GcEventMonitor {
       ExecutorService executorService,
       GcEventMonitorConfiguration gcEventMonitorConfiguration) {
     GcMonitorFactory gcMonitorFactory =
-        ProfilerServiceInitializer.findServiceLoader(GcMonitorFactory.class);
+        ProfilingInitializer.findServiceLoader(GcMonitorFactory.class);
 
     if (gcMonitorFactory != null) {
       init(
